culinarycodex/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
SQLAlchemy==2.0.32
python-dotenv==1.0.1
